feat: Use a new config structure
This PR is a breaking change and is incompatible with the previous
`config.yml` structure. `config.yml` is now rewired to have a new parent
key `workloads` which accepts an array to represent different
applications. This structure is more UX friendly and visualises all the
resources required to run the app in a more modular way.

Closes https://github.com/mr-karan/kubekutr/issues/2